refactor: 댓글리스트 fetching 동작 이동
페이지네이션 구현 위해 comment List fetching 동작 App.js로 이동
