+++ wy/src/App.js
@@ -1,12 +1,25 @@
-import React from 'react';
+import React, { useState, useEffect } from 'react';
 import CommentListContainer from './containers/CommentListContainer';
 import PageListContainer from './containers/PageListContainer';
 import FormContainer from './containers/FormContainer';
+import getCommentList from './api/getCommentList';
 
 function App() {
+
+  const [commentList, setCommentList] = useState([]);
+
+  async function fetchComments() {
+    const fetchedComments = await getCommentList();
+    setCommentList(fetchedComments);
+  }
+
+  useEffect(() => {
+    fetchComments();
+  });
+
   return (
     <div>
-      <CommentListContainer />
+      <CommentListContainer commentList={commentList}/>
       <PageListContainer />
       <FormContainer />
     </div>
