+++ wy/src/containers/CommentListContainer.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import CommentList from '../components/CommentList';
 
-function CommentListContainer() {
-  return <CommentList />;
+function CommentListContainer({commentList}) {
+  return <CommentList commentList={commentList}/>;
 }
 
 export default CommentListContainer;
